Added box-sizing:border-box for easier element layout
Adding box-sizing: border-box to all elements makes layout much simpler
as one can apply width: 100% and keep a padding for example. The padding
is not added on to the width.

More information on this can be found at
http://www.paulirish.com/2012/box-sizing-border-box-ftw/ and it's
generally considered safe to use on all elements.

From what I can see, this doesn't affect any existing content or
layouts.