Added footer area and ability to set title in footer
Instead of setting the title in the header bar, where space is limited,
I've added a footer overlay that mimics what's done on the photo and
album overlays.

This displays the title and in the future, the maybe the description.

This is a place where long titles and descriptions can be shown without
affecting other elements of the layout.