Detect accesses to admin functions by guests.
This commit detects accesses to admin functions when the user is a guest
and returns a corresponding error. This is now used to redirect the user
to the start page instead of "non-existing function". It has the
advantage that a bug in Lychee, that causes a non-existing function
to be called, can be spotted easier.